pyproject.toml
[tool.poetry]
name = "combine-images-to-pdf"
version = "0.1.0"
description = ""
authors = ["Wildiney Di Masi <wildiney@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.4.0"
reportlab = "^4.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
